+++ order-api/src/test/java/org/ecommerce/orderapi/controller/OrderControllerTest.java
@@ -11,6 +11,7 @@ import java.time.LocalDateTime;
 import java.util.List;
 
 import org.ecommerce.orderapi.dto.OrderDto;
+import org.ecommerce.orderapi.external.controller.OrderController;
 import org.ecommerce.orderapi.service.OrderService;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
